changed dynamic layer exchanger example
