--- examples/partial_weight_exchange_example/config.yaml
@@ -1,21 +1,0 @@
-# Parameters that describe server
-n_server_rounds: 4 # The number of rounds to run FL
-
-# Parameters that describe clients
-n_clients: 5 # The number of clients in the FL experiment
-local_epochs: 1 # The number of epochs to complete for client
-batch_size: 4 # The batch size for client training
-
-num_classes: 4 # number of classes in the classification task.
-
-sequence_length: 256 # length of tokenized sequences, must be >= 256.
-
-normalize: False # specifies whether division by the tensor's dimension is performed when computing its drift norm.
-filter_by_percentage: True # this argument toggles between the two mechanisms for selecting the tensors to be exchanged.
-select_drift_more: True # If True, then select those tensors that drift the most in l2 norm.
-
-norm_threshold: 24.5 # if a tensor's drift norm exceeds this threshold, then it will be sent to the server. Only applicable when filter_by_percentage is False
-exchange_percentage: 0.5 # if a tensor has a drift norm that is falls in the top percentile specified by this argument, then it will be sent to the server. Only applicable when filter_by_percentage is True
-
-sample_percentage: 0.75 # the proportion of the training data that will actually be used in training
-beta: 0.5 # controls the heterogeneity of the data distributions among clients
