--- examples/partial_weight_exchange_example/run_client.slrm
@@ -1,60 +1,0 @@
-#!/bin/bash
-
-#SBATCH --nodes=1
-#SBATCH --ntasks=1
-#SBATCH --ntasks-per-node=1
-#SBATCH --cpus-per-task=1
-#SBATCH --gres=gpu:1
-#SBATCH --mem=32G
-#SBATCH --partition=a40
-#SBATCH --qos=normal
-
-# Note:
-#	  ntasks: Total number of processes to use across world
-#	  ntasks-per-node: How many processes each node should create
-
-# Set NCCL options
-# export NCCL_DEBUG=INFO
-# NCCL backend to communicate between GPU workers is not provided in vector's cluster.
-# Disable this option in slurm.
-export NCCL_IB_DISABLE=1
-
-if [[ "${SLURM_JOB_PARTITION}" == "t4v2" ]] || \
-    [[ "${SLURM_JOB_PARTITION}" == "rtx6000" ]]; then
-    echo export NCCL_SOCKET_IFNAME=bond0 on "${SLURM_JOB_PARTITION}"
-    export NCCL_SOCKET_IFNAME=bond0
-fi
-
-# Process inputs
-
-SERVER_ADDRESS=$1
-PERCENTAGE=$2
-LOG_DIR=$3
-VENV_PATH=$4
-JOB_HASH=$5
-
-# Print relevant information for the client to connect to the server and run
-echo "Server Address: ${SERVER_ADDRESS}"
-echo "Python Venv Path: ${VENV_PATH}"
-echo "Job Hash: ${JOB_HASH}"
-
-LOG_DIR_FINAL="${LOG_DIR}_${PERCENTAGE}/"
-
-mkdir -p ${LOG_DIR_FINAL}
-
-LOG_PATH="${LOG_DIR_FINAL}client_log_${JOB_HASH}.log"
-
-echo "Placing logs in: ${LOG_DIR_FINAL}"
-echo "World size: ${SLURM_NTASKS}"
-echo "Number of nodes: ${SLURM_NNODES}"
-NUM_GPUs=$(nvidia-smi --query-gpu=name --format=csv,noheader | wc -l)
-echo "GPUs per node: ${NUM_GPUs}"
-
-# Source the environment
-source ${VENV_PATH}bin/activate
-echo "Active Environment:"
-which python
-
-echo "Server Address used by Client: ${SERVER_ADDRESS}"
-
-python -m examples.partial_weight_exchange_example.client --server_address ${SERVER_ADDRESS} > ${LOG_PATH} 2>&1
